src/AIGCBrowserSrc/ChromiumSrcPatch/third_party/blink/web_tests/fast/loader/target-blank-to-local-resource.html

<a id="a" target="_blank" href="tangram://downloads"></a>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.setPopupBlockingEnabled(false);
}
// The navigation should be blocked before a new window is opened.
a.click();
</script>
